Connection test page (#382)
This is temporary - we have this page live now and don't want it to go
away when we deploy to live. A better integrated option would be
nice-to-have for the future.


The second part of this test isn't really relevant once we use the new simx approach. But leaving as it is for now.