Renames and common structures (#439)
Rename using action in component names (I'll take this further separately) and using terms that correspond better with the UI we see.

Remove GridItem from components, preferring to add them in the Grid, to give more naturally named and potentially reusable components. This might make it easier to try other layouts and in any case makes it easier to understand the one we have.