Context-aware tour (#486)
Adds the ability to re-trigger a tour via a Help menu action,

To reduce the number of scenarios we require you to be connected to trigger the tour (prompted via dialog).

We show an adapted version of the "first data sample" tour steps in addition to the usual post-connect steps if you already have data samples when you connect.

Post-bulk-recording tours work again with a minor copy tweak.

I've not tried to do a specialised tour for data import / "Open in micro:bit CreateAI" etc. so if you don't connect you do miss out.